changed forms & inputs of login, register & shelter page
